test(cli): close full-suite regressions
