+++ docs/cli-fixtures/help-layout-v1.json
@@ -4,7 +4,7 @@
   "summaries": {
     "solve": "Solve a required plain-language goal; defaults to 5 iterations. Next: show.",
     "run": "Run a scenario (required unless configured); defaults to 1 iteration. Next: status.",
-    "status": "Show a required run's current snapshot; defaults to text output. Next: watch.",
+    "status": "Show a required run's current snapshot; defaults to text output. Queue: autoctx queue status. Next: watch.",
     "watch": "Follow a required run; defaults to polling every 2 seconds. Next: show.",
     "show": "Inspect a required run; defaults to its latest generation. Next: export.",
     "export": "Export a required run or scenario; defaults to JSON on stdout. Next: import-package."
