+++ ts/src/cli/cli-contract.ts
@@ -136,30 +136,36 @@ export function loadContract(path: string): Contract {
 }
 
 function validateVersion2Entries(value: unknown): void {
-  if (!value || typeof value !== "object" || Array.isArray(value)) return;
-  const record = value as Record<string, unknown>;
+  const record = objectRecord(value);
+  if (!record) return;
   if (typeof record.schema_version !== "number" || record.schema_version < 2) return;
   if (!Array.isArray(record.commands)) return;
   const required = ["positionals", "flags", "output", "exit_codes", "examples", "runtime_shapes"];
   for (const item of record.commands) {
-    if (!item || typeof item !== "object" || Array.isArray(item)) continue;
-    const command = item as Record<string, unknown>;
+    const command = objectRecord(item);
+    if (!command) continue;
     const missing = required.filter((field) => !(field in command));
     if (missing.length > 0) {
       throw new Error(
         `contract v2 command ${JSON.stringify(command.id ?? "<unknown>")} missing fields: ${missing.join(", ")}`,
       );
     }
-    const support = command.runtime_support as Record<string, { status?: unknown }> | undefined;
-    const shapes = command.runtime_shapes as Record<string, unknown> | undefined;
+    const support = objectRecord(command.runtime_support);
+    const shapes = objectRecord(command.runtime_shapes);
     for (const runtime of ["python", "typescript"] as const) {
-      if (support?.[runtime]?.status === "yes" && shapes?.[runtime] === undefined) {
+      const runtimeSupport = objectRecord(support?.[runtime]);
+      if (runtimeSupport?.status === "yes" && shapes?.[runtime] === undefined) {
         throw new Error(
           `contract v2 command ${JSON.stringify(command.id ?? "<unknown>")} missing ${runtime} runtime shape`,
         );
       }
     }
   }
+}
+
+function objectRecord(value: unknown): Record<string, unknown> | undefined {
+  const parsed = z.record(z.unknown()).safeParse(value);
+  return parsed.success ? parsed.data : undefined;
 }
 
 /** Return the canonical command id every alias resolves to, or
